requirements.txt
fastapi==0.116.1
starlette==0.47.2
uvicorn[standard]==0.30.6
langserve==0.3.1
langchain==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.31
openai==1.101.0
tiktoken==0.7.0
qdrant-client==1.15.1
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
httpx==0.28.1
sse-starlette==2.1.2
pypdf==6.0.0
python-docx==1.1.2
